requirements.txt
bs4==0.0.1
gcloud==0.18.3
lxml==4.6.2
pycrypto==2.6.1
python-jwt==3.3.0
sseclient==0.0.27
firebase==3.0.1
pymongo==3.11.3
dnspython==2.1.0
requests-toolbelt==0.8.0
numpy==1.19.4